+++ primefaces/src/main/resources/META-INF/resources/primefaces/tabview/tabview.js
@@ -521,19 +521,29 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
      */
     show: function(newPanel) {
         var headers = this.headerContainer,
+        actionsQuery = '.ui-tabs-actions:not(.ui-tabs-actions-global)',
         oldHeader = headers.filter('.ui-state-active'),
-        oldActions = oldHeader.next('.ui-tabs-actions'),
+        oldActions = oldHeader.next(actionsQuery),
+        hasOldActions = oldActions.length,
         newHeader = headers.eq(newPanel.index()),
-        newActions = newHeader.next('.ui-tabs-actions'),
+        newActions = newHeader.next(actionsQuery),
+        hasNewActions = newActions.length,
+        globalActions = this.navContainer.children('.ui-tabs-actions.ui-tabs-actions-global'),
         oldPanel = this.panelContainer.children('.ui-tabs-panel:visible'),
         $this = this;
+
+        globalActions.hide();
+        if(!hasNewActions){
+            newActions = globalActions;
+            hasNewActions = newActions.length;
+        }
 
         //aria
         oldPanel.attr('aria-hidden', true);
         oldPanel.addClass('ui-helper-hidden');
         oldHeader.attr('aria-expanded', false);
         oldHeader.attr('aria-selected', false);
-        if(oldActions.length != 0) {
+        if(hasOldActions) {
             oldActions.attr('aria-hidden', true);
         }
 
@@ -541,14 +551,14 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
         newPanel.removeClass('ui-helper-hidden');
         newHeader.attr('aria-expanded', true);
         newHeader.attr('aria-selected', true);
-        if(newActions.length != 0) {
+        if(hasNewActions) {
             newActions.attr('aria-hidden', false);
         }
 
         if(this.cfg.effect && oldPanel.length) {
             oldPanel.hide(this.cfg.effect, null, this.cfg.effectDuration, function() {
                 oldHeader.removeClass('ui-tabs-selected ui-state-active');
-                if(oldActions.length != 0) {
+                if(hasOldActions) {
                     oldActions.hide($this.cfg.effect, null, $this.cfg.effectDuration);
                 }
 
@@ -556,7 +566,7 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
                 newPanel.show($this.cfg.effect, null, $this.cfg.effectDuration, function() {
                     $this.postTabShow(newPanel);
                 });
-                if(newActions.length != 0) {
+                if(hasNewActions) {
                     newActions.show($this.cfg.effect, null, $this.cfg.effectDuration);
                 }
             });
@@ -564,13 +574,13 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
         else {
             oldHeader.removeClass('ui-tabs-selected ui-state-active');
             oldPanel.hide();
-            if(oldActions.length != 0) {
+            if(hasOldActions) {
                 oldActions.hide();
             }
 
             newHeader.addClass('ui-tabs-selected ui-state-active');
             newPanel.show();
-            if(newActions.length != 0) {
+            if(hasNewActions) {
                 newActions.show();
             }